requirements.txt
Babel==2.4.0

# Django
Django==1.11.4
django-localflavor==1.5.2
django-phonenumber-field==1.3.0
djangorestframework===3.6.3
django-enumchoicefield==0.9.0

#
olefile==0.44
phonenumberslite==8.7.1

#
Pillow==4.2.1
pkg-resources==0.0.0
psycopg2==2.7.3
pytz==2017.2
